mgr/dashboard: refactor pool info retrieval to service class
